added clhlock armv7. Proper casting of tid

+++ benchmarks/ck/clhlock.c
@@ -30,8 +30,8 @@ void *run(void *arg)
 int main()
 {
     pthread_t threads[NTHREADS];
-    long tids[NTHREADS];
-    long i;
+    int tids[NTHREADS];
+    int i;
 
     ck_spinlock_clh_t unowned;
     ck_spinlock_clh_init(&lock, &unowned);
